Update README.md
This is my Git repository for a project in SDEV140. 
This is a quiz/trivia game that will be called Triviamaniac
there are separate topics within the game that the user can select
The game will track the user's points and offer a bonus question if all questions answered were correct.